+++ scripts/run_offline_regression.sh
@@ -141,6 +141,7 @@ swiftc \
   "$CORE/Models/SampleReceipt.swift" \
   "$CORE/Models/MonthlySnapshot.swift" \
   "$CORE/Models/TodaySpendingSummary.swift" \
+  "$CORE/Models/SyncMetadata.swift" \
   "$CORE/Models/Transaction.swift" \
   "$CORE/Models/Subscription.swift" \
   "$CORE/Models/BackupBundle.swift" \